정건효 : 레이 캐스팅, 지형 피킹_마우스로 지형 클릭해서 테셀레이션 높이맵 적용

+++ Project/Engine/CResMgr_init.cpp
@@ -406,7 +406,7 @@ void CResMgr::CreateDefaultMesh()
 	}
 
 	pMesh = new CMesh(true);
-	pMesh->Create(arrCube, 24, vecIdx.data(), vecIdx.size());
+	pMesh->Create(arrCube, 24, vecIdx.data(), (UINT)vecIdx.size());
 	AddRes<CMesh>(L"CubeMesh", pMesh);
 	vecIdx.clear();
 
@@ -516,7 +516,7 @@ void CResMgr::CreateDefaultMesh()
 	}
 
 	pMesh = new CMesh(true);
-	pMesh->Create(vecVtx.data(), vecVtx.size(), vecIdx.data(), vecIdx.size());
+	pMesh->Create(vecVtx.data(), (UINT)vecVtx.size(), vecIdx.data(), (UINT)vecIdx.size());
 	AddRes<CMesh>(L"SphereMesh", pMesh);
 	vecVtx.clear();
 	vecIdx.clear();
